frontend/bot_frontend/pyproject.toml
[project]
name = "bot_frontend"
version = "0.1.0"
description = "Telegram bot frontend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.10",
    "httpx>=0.27.0",
    "pydantic-settings>=2.7.1",
    "exams",
    "lectures",
    "rooms",
]

[tool.uv.sources]
exams = { workspace = true }
lectures = { workspace = true }
rooms = { workspace = true }

[project.scripts]
bot_frontend = "bot_frontend.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
extend = "../../pyproject.toml"
